Implemented the check initial controller and check InitialMeansAssessmentService classes. Refactor some of the existing to handle the flow from the rest controller.
