+++ crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/service/InitialMeansAssessmentService.java
@@ -1,51 +1,143 @@
 package uk.gov.justice.laa.crime.meansassessment.service;
 
+import io.sentry.Sentry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
-import uk.gov.justice.laa.crime.meansassessment.client.AuthorisationMeansAssessmentClient;
-import uk.gov.justice.laa.crime.meansassessment.exception.AssessmentCriteriaNotFoundException;
-import uk.gov.justice.laa.crime.meansassessment.exception.MeansAssessmentValidationException;
-import uk.gov.justice.laa.crime.meansassessment.model.AuthorizationResponse;
-import uk.gov.justice.laa.crime.meansassessment.model.common.ApiCreateMeansAssessmentRequest;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.client.WebClient;
+import uk.gov.justice.laa.crime.meansassessment.builder.CreateInitialAssessmentBuilder;
+import uk.gov.justice.laa.crime.meansassessment.config.MaatApiConfiguration;
+import uk.gov.justice.laa.crime.meansassessment.dto.InitialMeansAssessmentDTO;
+import uk.gov.justice.laa.crime.meansassessment.exception.APIClientException;
+import uk.gov.justice.laa.crime.meansassessment.model.common.*;
 import uk.gov.justice.laa.crime.meansassessment.staticdata.entity.AssessmentCriteriaEntity;
-import uk.gov.justice.laa.crime.meansassessment.staticdata.repository.AssessmentCriteriaRepository;
-import uk.gov.justice.laa.crime.meansassessment.validation.service.MeansAssessmentValidationService;
-import uk.gov.justice.laa.crime.meansassessment.validation.validator.CreateAssessmentValidator;
+import uk.gov.justice.laa.crime.meansassessment.staticdata.enums.CaseType;
+import uk.gov.justice.laa.crime.meansassessment.staticdata.enums.CurrentStatus;
+import uk.gov.justice.laa.crime.meansassessment.staticdata.enums.InitialAssessmentResult;
 
-import java.time.LocalDateTime;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.List;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class InitialMeansAssessmentService {
 
-    private final CreateAssessmentValidator createAssessmentValidator;
-    private final AssessmentCriteriaRepository assessmentCriteriaRepository;
+    @Qualifier("maatAPIOAuth2WebClient")
+    private final WebClient webClient;
+    private final MaatApiConfiguration configuration;
+    private final AssessmentCriteriaService assessmentCriteriaService;
+    private final CreateInitialAssessmentBuilder createInitialAssessmentBuilder;
+    private final AssessmentCriteriaChildWeightingService childWeightingService;
 
-    public void createInitialAssessment(ApiCreateMeansAssessmentRequest apiCreateMeansAssessmentRequest) throws AssessmentCriteriaNotFoundException {
+    public ApiCreateMeansAssessmentResponse createInitialAssessment(ApiCreateMeansAssessmentRequest meansAssessment) {
+        log.info("Create initial means assessment - Start");
+        List<ApiAssessmentSectionSummary> sectionSummaries = meansAssessment.getSectionSummaries();
+        AssessmentCriteriaEntity assessmentCriteria =
+                assessmentCriteriaService.getAssessmentCriteria(
+                        meansAssessment.getAssessmentDate(), meansAssessment.getHasPartner(), meansAssessment.getPartnerContraryInterest()
+                );
 
-        log.info("Starting initial means assessment");
-        createAssessmentValidator.validate(apiCreateMeansAssessmentRequest);
-        log.info("Validation completed for Rep ID {}", apiCreateMeansAssessmentRequest.getRepId());
-        getAssessmentCriteria(apiCreateMeansAssessmentRequest.getAssessmentDate(),
-                apiCreateMeansAssessmentRequest.getHasPartner(),
-                apiCreateMeansAssessmentRequest.getPartnerContraryInterest());
+        BigDecimal annualTotal = getAnnualTotal(meansAssessment.getCaseType(), assessmentCriteria, sectionSummaries);
+        BigDecimal adjustedIncomeValue = getAdjustedIncome(meansAssessment, assessmentCriteria, annualTotal);
 
+        InitialAssessmentResult result;
+        CurrentStatus status = meansAssessment.getAssessmentStatus();
+        String newWorkReasonCode = meansAssessment.getNewWorkReason().getCode();
+        if (status.equals(CurrentStatus.COMPLETE)) {
+            result = getAssessmentResult(adjustedIncomeValue, assessmentCriteria, newWorkReasonCode);
+        } else {
+            result = InitialAssessmentResult.NONE;
+        }
+
+        log.info("Initial means assessment calculation complete for Rep ID: {}", meansAssessment.getRepId());
+
+        ApiCreateAssessment assessment = createInitialAssessmentBuilder.build(
+                new InitialMeansAssessmentDTO(annualTotal, status, adjustedIncomeValue, result, assessmentCriteria, meansAssessment, sectionSummaries));
+        return persistAssessment(assessment, meansAssessment.getLaaTransactionId());
     }
 
-    private List<AssessmentCriteriaEntity> getAssessmentCriteria(LocalDateTime assessmentDate, boolean hasPartner, boolean contraryInterest) throws AssessmentCriteriaNotFoundException {
-        List<AssessmentCriteriaEntity> assessmentCriteriaForDate = assessmentCriteriaRepository.findAssessmentCriteriaForDate(assessmentDate);
-        if(!assessmentCriteriaForDate.isEmpty()){
-            // If there is no partner or there is a partner with contrary interest, set partnerWeightingFactor to null
-            if(!hasPartner ||  contraryInterest){
-                assessmentCriteriaForDate.forEach(ac -> ac.setPartnerWeightingFactor(null));
-            }
-            return assessmentCriteriaForDate;
-        } else {
-            log.error("No Assessment Criteria found for date {}", assessmentDate);
-            throw new AssessmentCriteriaNotFoundException(String.format("No Assessment Criteria found for date %s",assessmentDate));
+    protected BigDecimal getAdjustedIncome(ApiCreateMeansAssessmentRequest meansAssessment, AssessmentCriteriaEntity assessmentCriteria, BigDecimal annualTotal) {
+        BigDecimal totalChildWeighting =
+                childWeightingService.getTotalChildWeighting(meansAssessment.getChildWeightings(), assessmentCriteria);
+
+        if (BigDecimal.ZERO.compareTo(annualTotal) <= 0) {
+            return annualTotal
+                    .divide(assessmentCriteria.getApplicantWeightingFactor()
+                                    .add(assessmentCriteria.getPartnerWeightingFactor())
+                                    .add(totalChildWeighting),
+                            RoundingMode.UP);
         }
+        return BigDecimal.ZERO;
+    }
+
+    protected InitialAssessmentResult getAssessmentResult(BigDecimal adjustedIncomeValue, AssessmentCriteriaEntity assessmentCriteria, String newWorkReasonCode) {
+        if (adjustedIncomeValue.compareTo(assessmentCriteria.getInitialLowerThreshold()) <= 0) {
+            return InitialAssessmentResult.PASS;
+        } else if (adjustedIncomeValue.compareTo(assessmentCriteria.getInitialUpperThreshold()) >= 0) {
+            // TODO: Comment in PL/SQL suggests this should also apply to crown court cases
+            if (newWorkReasonCode.equalsIgnoreCase("HR")) {
+                return InitialAssessmentResult.HARDSHIP;
+            } else {
+                return InitialAssessmentResult.FAIL;
+            }
+        } else {
+            return InitialAssessmentResult.FULL;
+        }
+    }
+
+    protected BigDecimal getAnnualTotal(CaseType caseType, AssessmentCriteriaEntity assessmentCriteria, List<ApiAssessmentSectionSummary> sectionSummaries) {
+        BigDecimal annualTotal = BigDecimal.ZERO;
+        for (ApiAssessmentSectionSummary sectionSummary : sectionSummaries) {
+            for (ApiAssessmentDetail assessmentDetail : sectionSummary.getAssessmentDetails()) {
+                assessmentCriteriaService.checkAssessmentDetail(caseType, sectionSummary.getSection(), assessmentCriteria, assessmentDetail);
+                annualTotal = annualTotal.add(getDetailTotal(assessmentDetail));
+            }
+        }
+        return annualTotal;
+    }
+
+    protected BigDecimal getDetailTotal(ApiAssessmentDetail assessmentDetail) {
+        BigDecimal detailTotal = BigDecimal.ZERO;
+        BigDecimal partnerAmount = assessmentDetail.getPartnerAmount();
+        if (partnerAmount != null && !BigDecimal.ZERO.equals(partnerAmount)) {
+            detailTotal = detailTotal.add(
+                    partnerAmount.multiply(
+                            BigDecimal.valueOf(assessmentDetail.getPartnerFrequency().getWeighting())
+                    )
+            );
+        }
+        BigDecimal applicationAmount = assessmentDetail.getApplicantAmount();
+        if (applicationAmount != null && !BigDecimal.ZERO.equals(applicationAmount)) {
+            detailTotal = detailTotal.add(
+                    applicationAmount.multiply(
+                            BigDecimal.valueOf(assessmentDetail.getApplicantFrequency().getWeighting())
+                    )
+            );
+        }
+        return detailTotal;
+    }
+
+    public ApiCreateMeansAssessmentResponse persistAssessment(ApiCreateAssessment createAssessment, String laaTransactionId) {
+        ApiCreateMeansAssessmentResponse response = webClient.post()
+                .uri(configuration.getFinancialAssessmentEndpoints().getCreateUrl())
+                .headers(httpHeaders -> httpHeaders.setAll(Map.of(
+                        "Laa-Transaction-Id", laaTransactionId
+                )))
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(BodyInserters.fromValue(createAssessment))
+                .retrieve()
+                .bodyToMono(ApiCreateMeansAssessmentResponse.class)
+                .onErrorMap(throwable -> new APIClientException("Call to Court Data API failed, invalid response."))
+                .doOnError(Sentry::captureException)
+                .block();
+
+        log.info(String.format("Response from Court Data API: %s", response));
+        return response;
     }
 }
