+++ crime-means-assessment/src/main/resources/schemas/common/apiContribution.json
@@ -1,0 +1,41 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "id": "apiContribution.json",
+  "type": "object",
+  "description": "Crown court contributions",
+  "properties": {
+    "id": {
+      "type": "integer",
+      "title": "contribution unique id"
+    },
+    "monthly": {
+      "type": "number",
+      "title": "crown court rep type"
+    },
+    "upfront": {
+      "type": "number",
+      "title": "crown court rep order date"
+    },
+    "effectiveDate": {
+      "type": "string",
+      "title": "sentence order date"
+    },
+    "calculationDate": {
+      "type": "string",
+      "title": "Calculation date"
+    },
+    "capped": {
+      "type": "number",
+      "title": "maximum capped value"
+    },
+    "upliftApplied": {
+      "type": "boolean",
+      "description": "uplifted applied"
+    },
+    "basedOn": {
+      "type": "string",
+      "description": "contributions based on"
+    }
+  },
+  "additionalProperties": false
+}
