+++ crime-means-assessment/src/main/resources/schemas/common/apiCrownCourtSummary.json
@@ -1,0 +1,40 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "id": "apiCrownCourtOverview.json",
+  "type": "object",
+  "description": "Crown Court Overview",
+  "properties": {
+    "ccRepId": {
+      "type": "number",
+      "title": "crown court rep Id"
+    },
+    "repType": {
+      "type": "string",
+      "title": "crown court rep type"
+    },
+    "repOrderDate": {
+      "type": "string",
+      "title": "crown court rep order date"
+    },
+    "sentenceOrderDate": {
+      "type": "string",
+      "title": "sentence order date"
+    },
+    "withdrawalDate": {
+      "type": "string",
+      "title": "Crown Court withdrawal date"
+    },
+    "repOrderDecision": {
+      "type": "string",
+      "title": " rep order decision"
+    },
+    "inPrisoned": {
+      "type": "boolean",
+      "description": "If in prison"
+    }
+  },
+  "additionalProperties": false,
+  "required": [
+    "repOrderDecision"
+  ]
+}
